+++ src/inventory/dto/create-inventory.dto.ts
@@ -1,5 +1,5 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { IsArray, IsMongoId, IsNotEmpty, IsOptional, IsString } from 'class-validator';
+import { IsArray, IsEnum, IsMongoId, IsNotEmpty, IsOptional, IsString } from 'class-validator';
 
 export class CreateInventoryDto {
   @ApiProperty({ example: 'Canon mf3010' })
@@ -17,9 +17,15 @@ export class CreateInventoryDto {
   @IsString()
   serial?: string;
 
-  @ApiProperty()
+  @ApiProperty({ required: false, enum: ['active', 'repair', 'broken', 'inactive'] })
+  @IsOptional()
+  @IsEnum(['active', 'repair', 'broken', 'inactive'])
+  status?: string;
+
+  @ApiProperty({ required: false })
   @IsMongoId()
-  user: string;
+  @IsOptional()
+  user?: string;
 
   @ApiProperty({ required: false })
   @IsOptional()