+++ src/inventory/dto/update-inventory.dto.ts
@@ -3,8 +3,8 @@ import { IsEnum, IsOptional } from 'class-validator';
 import { CreateInventoryDto } from './create-inventory.dto';
 
 export class UpdateInventoryDto extends PartialType(CreateInventoryDto) {
-  @ApiProperty({ required: false, enum: ['active', 'repair', 'broken'] })
+  @ApiProperty({ required: false, enum: ['active', 'repair', 'broken', 'inactive'] })
   @IsOptional()
-  @IsEnum(['active', 'repair', 'broken'])
+  @IsEnum(['active', 'repair', 'broken', 'inactive'])
   status?: string;
 }