+++ src/inventory/inventory.module.ts
@@ -18,7 +18,7 @@ import { CommonModule } from '../common/common.module';
     ConfigModule,
     MulterModule.register({
       storage: require('multer').diskStorage({
-        destination: './uploads',
+        destination: './uploads/inventory',
         filename: (req, file, cb) => {
           const ext = file.originalname.split('.').pop();
           const name = `${Date.now()}-${Math.round(Math.random() * 1e9)}.${ext}`;